moved Address to top level types file. fixed circular deps with asset-swapper/types and asset-swapper/token_adjacency_graph
